src/temperature_humid_monitor/src/requirements.txt
influxdb-client==1.36.0
paho-mqtt==2.0.0

